% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag.transform.R
\name{ag.transform}
\alias{ag.transform}
\title{Transform data to Andersen-Gill format}
\usage{
ag.transform(dataset, id, age, event)
}
\arguments{
\item{dataset}{the data frame of interest}

\item{id}{the variable that uniquely identifies each unit}

\item{age}{the variable that give the "age" of the unit}

\item{event}{the event indicator; 0/1 == nonevent/event}
}
\value{
The same data frame with three variables (columns) added: \code{time1},
    \code{time2}, \code{AG_event}; and one variable removed: \code{event}.
}
\description{
\code{ag.transform} reformats your data frame to the Andersen-Gill style, which includes the triplet (\code{time1}, \code{time2}, \code{event}); and deletes the former \code{event} variable.
}
\examples{
ag_style <- ag.transform(thermoking_sample_raw, serial, months_in_field, event)
all.equal(thermoking_sample, ag_style)
}
