% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thermoking_sample}
\alias{thermoking_sample}
\title{Thermo King data, 100 random units, Andersen-Gill style}
\format{A data frame with 400 rows and 87 variables; notable variables:
\describe{
  \item{months_in_field}{number of months in the field for the unit.}
  \item{serial}{unit identifier.}
  \item{time1}{start of time interval; first of A-G triplet.}
  \item{time2}{stop of time interval; second of A-G triplet.}
  \item{AG_event}{Indicator variable for last full month of survival; third of A-G triplet.}
  ...
}}
\source{
See source code \code{data-raw/LoadAndProcess.R}
}
\usage{
thermoking_sample
}
\description{
A sample of 100 units from \code{thermoking}, the Andersen-Gill formatted package dataset containing the time-varying covariates and survival information for Thermo King units. See \code{data-raw/LoadAndProcess.R} for processing code.
#'
}
\keyword{datasets}
