% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{thermoking_sample_raw}
\alias{thermoking_sample_raw}
\title{Thermo King data, 100 random units}
\format{A data frame with 500 rows and 85 variables; notable variables:
\describe{
  \item{months_in_field}{number of months in the field for the unit.}
  \item{serial}{unit identifier.}
  \item{event}{event indicator for month of event (unit failure); one month later than \code{AG_event}}
  ...
}}
\source{
See source code \code{data-raw/LoadAndProcess.R}
}
\usage{
thermoking_sample_raw
}
\description{
A sample of 100 units from the Thermo King data BEFORE modification to Andersen-Gill format; instead of (\code{time1}, \code{time2}, \code{AG_event}) triplet, it contains \code{months_in_field} and \code{event}. See \code{ag.transform()} example for transformation.
#'
}
\keyword{datasets}
